backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.7
cloudinary==1.38.0
colorama==0.4.6
dj-database-url==2.0.0
Django==4.2.1
django-cors-headers==3.14.0
django-environ==0.10.0
djangorestframework==3.14.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
Pillow==9.5.0
# psycopg2-binary==2.9.6
PyJWT==2.6.0
pytz==2023.3
requests==2.29.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uhashring==2.3
urllib3==1.26.15
uvicorn==0.23.2
whitenoise==6.6.0

